# Conflicts:
#	frontend/src/api/utils.ts
#	frontend/src/components/ui/Button.tsx
#	frontend/src/features/AnnotationFileRange/FileRangeActionBar.tsx
#	frontend/src/features/Annotator/Navigation/hooks.ts
#	frontend/src/features/Annotator/Skeleton.tsx
#	frontend/src/features/Annotator/Spectrogram/hooks.ts
#	frontend/src/features/Storage/display/Item.tsx
#	frontend/src/routes/_authenticated/_admin.tsx
#	frontend/src/routes/_authenticated/_admin/annotation-campaign/new.tsx
#	frontend/src/routes/_authenticated/_admin/dataset/$datasetID.tsx
#	frontend/src/routes/_authenticated/_admin/dataset/index.tsx
#	frontend/src/routes/_authenticated/_admin/storage.tsx
#	frontend/src/routes/_authenticated/_superuser.tsx
#	frontend/src/routes/_authenticated/_superuser/ontology/$type.$id.tsx
#	frontend/src/routes/_authenticated/annotation-campaign/$campaignID/_detailLayout/index.tsx
#	frontend/src/routes/_authenticated/annotation-campaign/$campaignID/_detailLayout/phase.$phaseType.tsx
#	frontend/src/routes/_authenticated/annotation-campaign/$campaignID/phase.$phaseType/edit-annotators.tsx
#	frontend/src/routes/_authenticated/annotation-campaign/$campaignID/phase.$phaseType/spectrogram/$spectrogramID.tsx
#	frontend/src/routes/_authenticated/annotation-campaign/$campaignID/route.tsx
#	frontend/src/routes/_authenticated/annotation-campaign/index.tsx
#	frontend/src/routes/_authenticated/route.tsx
#	frontend/tests/05-PhaseDetail.spec.ts
#	frontend/tests/utils/fixture.ts

+++ backend/api/schema/nodes/annotation_spectrogram.py
@@ -44,7 +44,7 @@ def get_task(
 
 class AnnotationSpectrogramNode(ExtendedNode):
 
-    duration = graphene.Int(required=True)
+    duration = graphene.Float(required=True)
     annotation_comments = AuthenticatedPaginationConnectionField(AnnotationCommentNode)
 
     class Meta:
